+++ src/view/common/ViewUtils.hpp
@@ -1,21 +1,24 @@
 #ifndef VIEWUTILS_HPP
 #define VIEWUTILS_HPP
 
+#include "view/elements/private/HsmElement.hpp"
+
 class QGraphicsScene;
 class QPointF;
 
 namespace view {
 
-class HsmElement;
-
 class ViewUtils {
 public:
+    // Need to specify childType if onlyAcceptsChildren is true
     static HsmElement* topHsmElementAt(QGraphicsScene* scene,
                                        const QPointF& scenePos,
                                        const bool onlyConnectable,
                                        const bool onlyAcceptsConnections,
                                        const bool onlyAcceptsChildren,
-                                       const HsmElement* ignoreElement);
+                                       const bool ignoreSelected,
+                                       const HsmElement* ignoreElement = nullptr,
+                                       const HsmElementType childType = HsmElementType::UNKNOWN);
 };
 
 };  // namespace view