+++ src/view/elements/HsmExitPointElement.cpp
@@ -14,6 +14,10 @@ bool HsmExitPointElement::isConnectable() const {
     return false;
 }
 
+bool HsmExitPointElement::canBeTopLevel() const {
+    return false;
+}
+
 void HsmExitPointElement::paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget) {
     HsmRoundElement::paint(painter, option, widget);