+++ src/view/elements/HsmExitPointElement.hpp
@@ -11,6 +11,7 @@ public:
     virtual ~HsmExitPointElement() = default;
 
     bool isConnectable() const override;
+    bool canBeTopLevel() const override;
 
 protected:
     void paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget) override;