+++ src/view/elements/HsmFinalElement.hpp
@@ -14,6 +14,7 @@ public:
     virtual ~HsmFinalElement() = default;
 
     bool isConnectable() const override;
+    bool canBeTopLevel() const override;
 
 protected:
     void paint(QPainter* painter, const QStyleOptionGraphicsItem* option, QWidget* widget) override;