+++ src/view/elements/HsmStateElement.hpp
@@ -22,7 +22,9 @@ public:
     virtual ~HsmStateElement() = default;
 
     void init(const QSharedPointer<model::StateMachineEntity>& modelEntity) override;
-    bool acceptsChildren() const override;
+    bool acceptsChildElement(const HsmElementType type) const override;
+    bool canBeTopLevel() const override;
+    
     virtual QList<QGraphicsItem*> hsmChildItems() const;
     bool hasSubstates() const override;
     // bool isDirectChild(HsmElement* item) const override;