+++ CHANGELOG.md
@@ -1,7 +1,16 @@
 # Changelog
 All notable changes to project will be documented in this file.
 
-## [0.2.0] - 2025-11-30
+## [0.2.0] - 2026-02-22
+### New
+- [REQ-103f1] History state must have a parent state
+- [REQ-103f2] Entry point must have a parent state
+- [REQ-103f3] Exit point must have a parent state
+- [REQ-103f4] Initial and final states shall be top level only
+- [REQ-103f5] Visual indication of restricted move action
+- [REQ-103f] All hierarchy restrictions implemented
+
+## [0.1.0] - 2025-11-30
 ### New
 - Rendering of transition callback and condition